+++ CHANGES.md
@@ -1,5 +1,4 @@
-## Changes in version 0.0.5 (in development)
-
+## Changes in version 0.0.5
 
 - In `s2gos_client`:
   - Improved error handling (#54)